Refactor: Consolidate VideoPlaybackManager and PlayerManager
- Consolidated PlayerManager functionality directly into VideoPlaybackManager
- Removed redundant delegation layer and async observation overhead
- Organized consolidated code into logical extensions:
  - VideoPlaybackManager+Creation.swift
  - VideoPlaybackManager+Playback.swift
  - VideoPlaybackManager+Monitoring.swift
- Updated tests to work with direct implementation
- Removed PlayerManager files and updated references
- Maintains exact same public API with improved performance

Resolves #53
