+++ LostArchiveTVTests/VideoPlaybackManagerTests.swift
@@ -24,32 +24,61 @@ struct VideoPlaybackManagerTests {
         
         // Assert
         #expect(playbackManager.player != nil)
+        #expect(playbackManager.currentVideoURL == url)
     }
     
     @Test
-    func play_setsPlayingStateToTrue() async throws {
+    func createNewPlayer_createsPlayerFromAsset() {
         // Arrange
         let playbackManager = VideoPlaybackManager()
         let url = URL(string: "https://example.com/test/test.mp4")!
         let asset = AVURLAsset(url: url)
-        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
-        playbackManager.useExistingPlayer(player)
         
-        // Act - directly modify underlying player manager
-        playbackManager.play()
-        
-        // We need to directly modify the isPlaying property for testing
-        // This reflects how the async property would be updated in reality
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
+        // Act
+        playbackManager.createNewPlayer(from: asset, url: url)
         
         // Assert
-        #expect(playbackManager.isPlaying)
+        #expect(playbackManager.player != nil)
+        #expect(playbackManager.currentVideoURL == url)
     }
     
     @Test
-    func cleanupPlayer_releasesPlayerResources() async throws {
+    func createNewPlayer_withStartPosition_seeksToPosition() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let startPosition = 10.0
+        
+        // Act
+        playbackManager.createNewPlayer(from: asset, url: url, startPosition: startPosition)
+        
+        // Assert
+        #expect(playbackManager.player != nil)
+        #expect(playbackManager.currentVideoURL == url)
+        // The player will seek to the start position, but we can't easily test the exact time
+        // since seeking is asynchronous
+    }
+    
+    @Test
+    func play_callsPlayerPlay() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        
+        // Act
+        playbackManager.play()
+        
+        // Assert
+        #expect(playbackManager.isPlaying == true)
+        #expect(playbackManager.player?.rate == 1.0)
+    }
+    
+    @Test
+    func pause_callsPlayerPause() {
         // Arrange
         let playbackManager = VideoPlaybackManager()
         let url = URL(string: "https://example.com/test/test.mp4")!
@@ -58,34 +87,61 @@ struct VideoPlaybackManagerTests {
         playbackManager.useExistingPlayer(player)
         playbackManager.play()
         
-        // Manually set the initial state for testing
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
+        // Act
+        playbackManager.pause()
+        
+        // Assert
+        #expect(playbackManager.isPlaying == false)
+        #expect(playbackManager.player?.rate == 0.0)
+    }
+    
+    @Test
+    func cleanupPlayer_releasesPlayerResources() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        playbackManager.play()
         
         // Initial state check
         #expect(playbackManager.player != nil)
         #expect(playbackManager.isPlaying)
+        #expect(playbackManager.currentVideoURL != nil)
         
         // Act
         playbackManager.cleanupPlayer()
         
-        // Manually set the final state for testing
-        await MainActor.run {
-            playbackManager.isPlaying = false
-            playbackManager.currentTime = 0
-            playbackManager.videoDuration = 0
-        }
-        
-        // Assert
+        // Assert - cleanupPlayer() directly sets all state
         #expect(playbackManager.player == nil)
-        #expect(!playbackManager.isPlaying)
+        #expect(playbackManager.isPlaying == false)
         #expect(playbackManager.currentTime == 0)
         #expect(playbackManager.videoDuration == 0)
+        #expect(playbackManager.currentVideoURL == nil)
     }
     
     @Test
-    func playerItemDidReachEnd_updatesPlayingState() async throws {
+    func seekToBeginning_seeksToZeroAndPlays() async {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        
+        // Act
+        playbackManager.seekToBeginning()
+        
+        // Wait a moment for the seek to complete
+        try? await Task.sleep(for: .milliseconds(100))
+        
+        // Assert - the implementation sets isPlaying to true in the completion handler
+        #expect(playbackManager.isPlaying == true)
+    }
+    
+    @Test
+    func setAndResetPlaybackRate_managesRateCorrectly() {
         // Arrange
         let playbackManager = VideoPlaybackManager()
         let url = URL(string: "https://example.com/test/test.mp4")!
@@ -94,10 +150,30 @@ struct VideoPlaybackManagerTests {
         playbackManager.useExistingPlayer(player)
         playbackManager.play()
         
-        // Manually set initial state for testing
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
+        let testRate: Float = 2.0
+        
+        // Act - set temporary rate
+        playbackManager.setTemporaryPlaybackRate(rate: testRate)
+        
+        // Assert temporary rate is set
+        #expect(playbackManager.player?.rate == testRate)
+        
+        // Act - reset rate
+        playbackManager.resetPlaybackRate()
+        
+        // Assert rate is reset (should be 1.0 as that's the normal rate)
+        #expect(playbackManager.player?.rate == 1.0)
+    }
+    
+    @Test
+    func playerItemDidReachEnd_restartsPlayback() async {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        playbackManager.play()
         
         // Initial state check
         #expect(playbackManager.isPlaying)
@@ -109,12 +185,37 @@ struct VideoPlaybackManagerTests {
             object: playerItem
         )
         
-        // Make sure isPlaying stays true in our test
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
+        // Wait a moment for the notification to be processed
+        try? await Task.sleep(for: .milliseconds(100))
         
         // Assert - the playback manager restarts the video and keeps playing
-        #expect(playbackManager.isPlaying)
+        #expect(playbackManager.isPlaying == true)
+        #expect(playbackManager.player?.rate == 1.0)
+    }
+    
+    @Test
+    func audioSessionSetup_configuresCorrectly() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        
+        // Act & Assert - should not throw
+        playbackManager.setupAudioSession()
+        playbackManager.setupAudioSession(forTrimming: true)
+        playbackManager.deactivateAudioSession()
+    }
+    
+    @Test
+    func computedProperties_returnCorrectValues() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        
+        // Assert
+        #expect(playbackManager.currentTimeAsCMTime != nil)
+        #expect(playbackManager.durationAsCMTime != nil)
+        #expect(playbackManager.currentVideoURL == url)
     }
 }
